Refactor code and move cookieListing to its component
