+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookieListing/cookieDetails/index.tsx
