+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookieListing/cookieDetails/stories/cookieDetails.stories.ts
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 /**
  * External dependencies.
  */
@@ -22,20 +21,19 @@ import type { Meta, StoryObj } from '@storybook/react';
 /**
  * Internal dependencies.
  */
-import { TempCookieData } from './tempData';
-import { CookieTable } from '../..';
+
+import CookieDetails from '..';
 
 const meta = {
-  title: 'Extension/DevTools/CookiesPanel/CookieTable',
-  component: CookieTable,
+  title: 'Extension/DevTools/CookiesPanel/CookieDetails',
+  component: CookieDetails,
   tags: ['autodocs'],
-} as Meta<typeof CookieTable>;
+} as Meta<typeof CookieDetails>;
 
 export default meta;
+
 type Story = StoryObj<typeof meta>;
 
 export const Primary: Story = {
-  args: {
-    cookies: TempCookieData,
-  },
+  args: {},
 };
