+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookieListing/cookieDetails/stories/details.stories.ts
