+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookieListing/cookieTable/index.tsx
