+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookieListing/cookieTable/stories/tempData.ts
